Change variable type

+++ include/Agent.hpp
@@ -44,7 +44,7 @@ namespace station_sim {
 		int gate_out{};
 		Point2D desired_location;
 		Point2D agent_location;
-		float steps_activate;
+		int steps_activate;
 		float wiggle;
 		float agent_max_speed{};
 		std::vector<float> agent_available_speeds;
